feat(actions): migrate upload init to Astro Action
Replace the /api/videos/upload REST endpoint with a new
actions.video.initUpload Astro Action. The TUS PATCH upload to
Cloudflare Stream stays client-side so progress tracking still works.

Closes #68
